Rama juampi final

+++ app/src/main/java/com/example/gimnasio_grupo3/fragments/Home.kt
@@ -1,5 +1,6 @@
 package com.example.gimnasio_grupo3.fragments
 
+import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
 import androidx.lifecycle.ViewModelProvider
@@ -10,16 +11,19 @@ import android.view.View
 import android.content.SharedPreferences
 import android.view.ViewGroup
 import android.widget.Button
+import android.widget.ImageView
 import android.widget.TextView
 import androidx.cardview.widget.CardView
 import androidx.navigation.findNavController
 import androidx.navigation.fragment.findNavController
+import com.example.gimnasio_grupo3.Firebase.FirebaseStorageConnection
 import com.example.gimnasio_grupo3.R
 import com.example.gimnasio_grupo3.activities.LoginActivity
 import com.example.gimnasio_grupo3.activities.MainActivity
 import com.example.gimnasio_grupo3.entities.Usuario
 import com.example.gimnasio_grupo3.sessions.MyPreferences
 import com.google.android.material.snackbar.Snackbar
+import retrofit2.Call
 
 class Home : Fragment() {
     lateinit var v : View
@@ -27,7 +31,9 @@ class Home : Fragment() {
     lateinit var btnUsuarios : Button
     lateinit var btnLogOut : Button
     lateinit var txtNombreCompleto : TextView
-    lateinit var txtTickets : TextView
+    lateinit var imgUsuario: ImageView
+    private var FirebaseStorageConnection = FirebaseStorageConnection()
+    lateinit var txtInfo : TextView
     private lateinit var myPreferences: MyPreferences
     private var user: Usuario? = null
     companion object {
@@ -46,18 +52,10 @@ class Home : Fragment() {
         btnLogOut = v.findViewById(R.id.btnLogOut)
 
         txtNombreCompleto = v.findViewById(R.id.txtNombreApellido)
-        txtTickets = v.findViewById(R.id.txtEmail3)
+        imgUsuario = v.findViewById(R.id.imgNavUsuario)
+        txtInfo = v.findViewById(R.id.txtEmail3)
         myPreferences = MyPreferences(requireContext())
         user = myPreferences.getUser()
-
-
-
-        btnLogOut.setOnClickListener {
-            val myPreferences = MyPreferences(requireContext())
-            myPreferences.deleteUser()
-            val intent = Intent(activity?.applicationContext, LoginActivity::class.java)
-            startActivity(intent)
-        }
 
         return v
     }
@@ -65,6 +63,7 @@ class Home : Fragment() {
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         viewModel = ViewModelProvider(this).get(HomeViewModel::class.java)
+        FirebaseStorageConnection.getImageLoadingAndImgDefault(imgUsuario,"usuarios/${user?.id}.jpg")
         // TODO: Use the ViewModel
     }
 
@@ -74,6 +73,7 @@ class Home : Fragment() {
         val myPreferences = MyPreferences(requireContext())
         val user = myPreferences.getUser()
 
+
         if (user != null) {
             setNombreCompleto(user.nombre, user.apellido)
             setTickets(user.ticketsRestantes.toString())
@@ -82,6 +82,12 @@ class Home : Fragment() {
         if (myPreferences.isAdmin() != true) {
             btnProfesores.visibility = View.GONE
             btnUsuarios.visibility = View.GONE
+        } else {
+            txtInfo.text = "Cuenta de administrador"
+            txtInfo.setCompoundDrawablesWithIntrinsicBounds(null, null, null, null)
+
+            val drawable = requireContext().getDrawable(R.drawable.setting)
+            txtInfo.setCompoundDrawablesWithIntrinsicBounds(drawable, null, null, null)
         }
 
         btnProfesores.setOnClickListener {
@@ -90,9 +96,15 @@ class Home : Fragment() {
         }
 
         btnLogOut.setOnClickListener {
-            myPreferences.deleteUser()
-            val intent = Intent(activity?.applicationContext, LoginActivity::class.java)
-            startActivity(intent)
+            confirmAction() { confirmed ->
+                if(confirmed){
+                    myPreferences.deleteUser()
+                    val intent = Intent(activity?.applicationContext, LoginActivity::class.java)
+                    startActivity(intent)
+                } else {
+                    Snackbar.make(v, "Acción cancelada", Snackbar.LENGTH_LONG).show()
+                }
+            }
         }
 
         btnUsuarios.setOnClickListener() {
@@ -109,13 +121,30 @@ class Home : Fragment() {
         }
     }
 
-    fun setNombreCompleto(nombre: String, apellido: String){
-        txtNombreCompleto.text = "${apellido}, ${nombre}"
+    private fun setNombreCompleto(nombre: String, apellido: String) {
+        val nombreCompleto = "${apellido}, ${nombre}"
+        txtNombreCompleto.text = nombreCompleto.uppercase()
     }
 
+    private fun setTickets(tickets: String){
+        txtInfo.text = "Tickets: ${tickets}"
+    }
 
+    private fun confirmAction(callback: (Boolean) -> Unit) {
+        val builder = AlertDialog.Builder(requireContext())
+        builder.setTitle("Cerrar Sesión")
+        builder.setMessage("¿Estás seguro de que deseas Cerrar Sesión?")
+        builder.setPositiveButton("Cerrar Sesión") { dialog, _ ->
+            callback(true) // Llama al callback con 'true' cuando el usuario confirma
+            dialog.dismiss()
+        }
 
-    fun setTickets(tickets: String){
-        txtTickets.text = "Tickets: ${tickets}"
+        builder.setNegativeButton("Cancelar") { dialog, _ ->
+            callback(false) // Llama al callback con 'false' cuando el usuario cancela
+            dialog.dismiss()
+        }
+
+        val dialog = builder.create()
+        dialog.show()
     }
 }
